Multi-Commit : Too many things:
	  Improved General shutter [TS1/TS2]
		Modified target system [allow modify objects]
		Wedge system for objects [grooves/cuts etc]
		Vec3D maths operators for the variable system
    SupplyPipe improvements
		TS1Upgrade changes.

+++ chipInc/ChipIRGuide.h
@@ -2,8 +2,8 @@
   CombLayer : MNCPX Input builder
  
  * File:   chipInc/ChipIRGuide.h
-*
- * Copyright (c) 2004-2013 by Stuart Ansell
+ *
+ * Copyright (c) 2004-2014 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -162,7 +162,8 @@ class ChipIRGuide : public attachSystem::TwinComp,
 
   /// Set surface
   void setMonoSurface(const int M) { monoWallSurf=M; }
-  void createAll(Simulation&,const shutterSystem::BulkShield&,const int);
+  void createAll(Simulation&,const shutterSystem::BulkShield&,
+		 const size_t);
 
   /// Access beam angle
   double getBeamAngle() const { return beamAngle; }
